backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
python-multipart==0.0.9    # Required for FastAPI file uploads
PyExifTool==0.5.6          # The standard for EXIF manipulation
google-cloud-vision==3.7.1 # Our OCR engine
sqlalchemy==2.0.28         # For PostgreSQL later
psycopg[binary]==3.3.3     # Postgres driver
python-dotenv==1.0.1       # To read our .env file